Inventory test 1
Inventory displays with button i. And the inventory displays on the player, and as you pick up items it will display in slot 1, then 2 then 3 as u keep picking up distinct items.

It also stacks the items until 99,

However only issue is that the item signal isnt detecting the player so it doesnt run the code to enter the inv.